+++ public/js/src/module/connection.js
@@ -3,10 +3,8 @@
  */
 
 import settings from './settings';
-
 import { t } from './translator';
 import utils from './utils';
-import $ from 'jquery';
 const parser = new DOMParser();
 const CONNECTION_URL = `${settings.basePath}/connection`;
 const TRANSFORM_URL = `${settings.basePath}/transform/xform${settings.enketoId ? `/${settings.enketoId}` : ''}`;
@@ -68,52 +66,58 @@ function uploadRecord( record ) {
  * @return { Promise }      [description]
  */
 function _uploadBatch( recordBatch ) {
-    return new Promise( ( resolve, reject ) => {
-        // Submission URL is dynamic, because settings.submissionParameter only gets populated after loading form from
-        // cache in offline mode.
-        const submissionUrl = ( settings.enketoId ) ? `${settings.basePath}/submission/${settings.enketoId}${_getQuery()}` : null;
+    // Submission URL is dynamic, because settings.submissionParameter only gets populated after loading form from
+    // cache in offline mode.
+    const submissionUrl = ( settings.enketoId ) ? `${settings.basePath}/submission/${settings.enketoId}${_getQuery()}` : null;
+    const controller = new AbortController();
 
-        $.ajax( submissionUrl, {
-            type: 'POST',
-            data: recordBatch.formData,
-            cache: false,
-            contentType: false,
-            processData: false,
-            headers: {
-                'X-OpenRosa-Version': '1.0',
-                'X-OpenRosa-Deprecated-Id': recordBatch.deprecatedId,
-                'X-OpenRosa-Instance-Id': recordBatch.instanceId
-            },
-            timeout: settings.timeout
-        } )
-            .done( ( data, textStatus, jqXHR ) => {
-                const result = {
-                    status: jqXHR.status,
-                    failedFiles: ( recordBatch.failedFiles ) ? recordBatch.failedFiles : undefined
-                };
-                if ( result.status === 201 || result.status === 202 ) {
-                    resolve( result );
-                } else {
-                    reject( result );
-                }
-            } )
-            .fail( jqXHR => {
-                let messageEl = null;
-                let message = null;
+    setTimeout( () => {
+        controller.abort();
+    }, settings.timeout );
+
+    return fetch( submissionUrl, {
+        method: 'POST',
+        cache: 'no-cache',
+        headers: {
+            'X-OpenRosa-Version': '1.0',
+            'X-OpenRosa-Deprecated-Id': recordBatch.deprecatedId,
+            'X-OpenRosa-Instance-Id': recordBatch.instanceId
+        },
+        signal: controller.signal,
+        body: recordBatch.formData
+    } )
+        .then( response => {
+            const result = {
+                status: response.status,
+                failedFiles: ( recordBatch.failedFiles ) ? recordBatch.failedFiles : undefined
+            };
+
+            if ( response.status === 400 ){
                 // 400 is a generic error. Any message returned by the server is probably more useful.
-                // Other more specific statusCodes will get harcoded and translated messages.
-                if ( jqXHR.status === 400 && jqXHR.responseXML ) {
-                    messageEl = jqXHR.responseXML.querySelector( 'OpenRosaResponse > message' );
-                    if ( messageEl ) {
-                        message = messageEl.textContent;
-                    }
-                }
-                reject( {
-                    status: jqXHR.status,
-                    message
-                } );
-            } );
-    } );
+                // Other more specific statusCodes will get hardcoded and translated messages.
+                return response.text()
+                    .then( text => {
+                        const xmlResponse = parser.parseFromString( text, 'text/xml' );
+                        if ( xmlResponse ){
+                            const messageEl = xmlResponse.querySelector( 'OpenRosaResponse > message' );
+                            if ( messageEl ) {
+                                result.message = messageEl.textContent;
+                            }
+                        }
+                        throw result;
+                    } );
+            } else if ( response.status !== 201  && response.status !== 202 ){
+                throw result;
+            } else {
+                return result;
+            }
+        } )
+        .catch( error => {
+            if ( error.name === 'AbortError' && typeof error.status === 'undefined' ){
+                error.status = 408;
+            }
+            throw error;
+        } );
 }
 
 /**
@@ -281,50 +285,90 @@ function getMaximumSubmissionSize( survey ) {
  * @return { Promise } a Promise that resolves with a form parts object
  */
 function getFormParts( props ) {
-    let error;
 
-    //TODO: use fetch
-    return new Promise( ( resolve, reject ) => {
-        $.ajax( TRANSFORM_URL + _getQuery(), {
-            type: 'POST',
-            data: {
-                serverUrl: props.serverUrl,
-                xformId: props.xformId,
-                xformUrl: props.xformUrl
-            }
-        } )
-            .done( data => {
-                data.enketoId = props.enketoId;
-                data.theme = data.theme || utils.getThemeFromFormStr( data.form ) || settings.defaultTheme;
-                _getExternalData( data )
-                    .then( resolve )
-                    .catch( reject );
-            } )
-            .fail( ( jqXHR, textStatus, errorMsg ) => {
-                if ( jqXHR.responseJSON && jqXHR.responseJSON.message && /ENOTFOUND/.test( jqXHR.responseJSON.message ) ) {
-                    jqXHR.responseJSON.message = 'Form could not be retrieved from server.';
+    return _postData( TRANSFORM_URL + _getQuery(), {
+        serverUrl: props.serverUrl,
+        xformId: props.xformId,
+        xformUrl: props.xformUrl
+    } )
+        .then( data => {
+            data.enketoId = props.enketoId;
+            data.theme = data.theme || utils.getThemeFromFormStr( data.form ) || settings.defaultTheme;
+
+            return _getExternalData( data );
+        } );
+}
+
+function _postData( url, data = {}  ){
+    return _request( url, 'POST', data );
+}
+
+function _getData( url, data = {} ){
+    return _request( url, 'GET', data );
+}
+
+function _request( url, method = 'POST', data = {}  ){
+    const options = {
+        method,
+        headers: { 'Content-Type': 'application/x-www-form-urlencoded','Accept': 'application/json' }
+    };
+    // add data
+    if ( method === 'GET' || method === 'HEAD' ){
+        if ( Object.keys( data ).length ){
+            const urlObj = new URL( url, location.href );
+            const search = urlObj.search.slice( 1 );
+            urlObj.search = `?${search}${search ? '&' : ''}${_encodeFormData( data )}`;
+            url = urlObj.href;
+        }
+    } else {
+        options.body = _encodeFormData( data );
+    }
+
+    return fetch( url, options )
+        .then( _throwResponseError )
+        .then( response => response.json() )
+        .catch(  data => {
+            const error = new Error( data.message );
+            error.status = data.status;
+            throw error;
+        } );
+}
+
+function _throwResponseError( response ){
+    if ( !response.ok ){
+        return response.json()
+            .then( data => {
+                if ( typeof data.status === 'undefined' ){
+                    data.status = response.status;
                 }
-                error = jqXHR.responseJSON || new Error( errorMsg );
-                error.status = jqXHR.status;
-                reject( error );
+                if ( typeof data.message === 'undefined' ){
+                    data.status = response.statusText;
+                }
+                throw data;
             } );
-    } );
+    } else {
+        return response;
+    }
+}
+
+function _encodeFormData( data ){
+    return Object.keys( data )
+        .filter( key => data[key] )
+        .map( key => encodeURIComponent( key ) + '=' + encodeURIComponent( data[key] ) )
+        .join( '&' );
 }
 
 function _getExternalData( survey ) {
-    let doc;
     const tasks = [];
 
     try {
-        // TODO: rewrite this
-        doc = $.parseXML( survey.model );
+        const doc = parser.parseFromString( survey.model, 'text/xml' );
 
-        survey.externalData = $( doc ).find( 'instance[id][src]' )
-            .map( ( index, el ) => ( {
-                id: $( el ).attr( 'id' ),
-                src: $( el ).attr( 'src' )
-            } ) ).get();
-        // end of rewrite TODO
+        survey.externalData = [ ...doc.querySelectorAll ( 'instance[id][src]' ) ]
+            .map( instance => ( {
+                id:  instance.id,
+                src: instance.getAttribute( 'src' )
+            } ) );
 
         survey.externalData
             .forEach( ( instance, index ) => {
@@ -336,8 +380,11 @@ function _getExternalData( survey ) {
                     } )
                     .catch( e => {
                         survey.externalData.splice( index, 1 );
-                        // let external data files fail quietly. Rely on Enketo Core to show error.
-                        console.error( e );
+                        // let external data files fail quietly in previews with ?form= parameter
+                        if ( !survey.enketoId ){
+                            return;
+                        }
+                        throw e;
                     } ) );
             } );
 
@@ -352,38 +399,23 @@ function _getExternalData( survey ) {
 
 /**
  * Obtains a media file
- * JQuery ajax doesn't support blob responses, so we're going native here.
  *
  * @param { string } url - a URL to a media file
  * @return {Promise<{url: string, item: Blob}>} a Promise that resolves with a media file object
  */
 function getMediaFile( url ) {
-    let error;
-    const xhr = new XMLHttpRequest();
 
-    // TODO: use fetch
-    return new Promise( ( resolve, reject ) => {
-        xhr.onreadystatechange = function() {
-            if ( this.readyState === 4 ) {
-                if ( this.status >= 200 && this.status < 300 ) {
-                    resolve( {
-                        url,
-                        item: this.response
-                    } );
-                } else {
-                    error = new Error( this.statusText || t( 'error.loadfailed', {
-                        resource: url
-                    } ) );
-                    error.status = this.status;
-                    reject( error );
-                }
-            }
-        };
-
-        xhr.open( 'GET', url );
-        xhr.responseType = 'blob';
-        xhr.send();
-    } );
+    return fetch( url )
+        .then( _throwResponseError )
+        .then( response =>  response.blob() )
+        .then( item => ( { url, item } ) )
+        .catch(  data => {
+            const error = new Error( data.message || t( 'error.loadfailed', {
+                resource: url
+            } ) );
+            error.status = data.status;
+            throw error;
+        } );
 }
 
 /**
@@ -409,11 +441,11 @@ function _getDataFile( url, languageMap ) {
                     result = utils.csvToXml( responseText, languageMap );
                     break;
                 case 'text/xml':
-                    result = ( new DOMParser() ).parseFromString( responseText, contentType );
+                    result = parser.parseFromString( responseText, contentType );
                     break;
                 default:
                     console.error( 'External data not served with expected Content-Type.', contentType );
-                    result = ( new DOMParser() ).parseFromString( responseText, 'text/xml' );
+                    result = parser.parseFromString( responseText, 'text/xml' );
             }
             if ( result && result.querySelector( 'parsererror' ) && contentType !== 'text/csv' ) {
                 console.log( 'Failed to parse external data as XML, am going to try as CSV' );
@@ -449,23 +481,10 @@ function getServiceWorkerVersion( serviceWorkerUrl ) {
         } );
 }
 
-// TODO: use fetch
 function getFormPartsHash() {
-    let error;
 
-    return new Promise( ( resolve, reject ) => {
-        $.ajax( TRANSFORM_HASH_URL + _getQuery(), {
-            type: 'POST'
-        } )
-            .done( data => {
-                resolve( data.hash );
-            } )
-            .fail( ( jqXHR, textStatus, errorMsg ) => {
-                error = new Error( errorMsg );
-                error.status = jqXHR.status;
-                reject( error );
-            } );
-    } );
+    return _postData( TRANSFORM_HASH_URL + _getQuery() )
+        .then( data => data.hash );
 }
 
 /**
@@ -475,21 +494,7 @@ function getFormPartsHash() {
  * @return { Promise<string> } a Promise that resolves with an XML instance as text
  */
 function getExistingInstance( props ) {
-    let error;
-
-    return new Promise( ( resolve, reject ) => {
-        $.ajax( INSTANCE_URL, {
-            type: 'GET',
-            data: props
-        } )
-            .done( data => {
-                resolve( data );
-            } )
-            .fail( ( jqXHR, textStatus, errorMsg ) => {
-                error = jqXHR.responseJSON || new Error( errorMsg );
-                reject( error );
-            } );
-    } );
+    return _getData( INSTANCE_URL, props );
 }
 
 // Note: settings.submissionParameter is only populated after loading form from cache in offline mode.